+++ persist/sqlite/hosts_test.go
@@ -9,14 +9,213 @@ import (
 	"go.sia.tech/coreutils/rhp/v4/siamux"
 	"lukechampine.com/frand"
 
+	proto2 "go.sia.tech/core/rhp/v2"
 	rhpv2 "go.sia.tech/core/rhp/v2"
+	proto3 "go.sia.tech/core/rhp/v3"
+	proto4 "go.sia.tech/core/rhp/v4"
 	rhpv4 "go.sia.tech/core/rhp/v4"
 	"go.sia.tech/core/types"
 	"go.sia.tech/coreutils/chain"
 	"go.sia.tech/explored/explorer"
 	"go.sia.tech/explored/geoip"
+	"go.sia.tech/explored/internal/testutil"
 	"go.uber.org/zap/zaptest"
 )
+
+func TestHostScan(t *testing.T) {
+	pk1 := types.GeneratePrivateKey()
+
+	n := newTestChain(t, false, nil)
+
+	assertHost := func(pubkey types.PublicKey, expected explorer.Host) {
+		hosts, err := n.db.QueryHosts(explorer.HostQuery{PublicKeys: []types.PublicKey{pubkey}}, explorer.HostSortPublicKey, explorer.HostSortAsc, 0, math.MaxInt64)
+		if err != nil {
+			t.Fatal(err)
+		}
+		testutil.Equal(t, "len(hosts)", 1, len(hosts))
+
+		testutil.Equal(t, "Host", expected, hosts[0])
+	}
+
+	// announce a host
+	const netAddr1 = "127.0.0.1:1234"
+	txn1 := types.Transaction{
+		ArbitraryData: [][]byte{
+			testutil.CreateAnnouncement(pk1, netAddr1),
+		},
+	}
+
+	n.mineTransactions(t, txn1)
+
+	hosts, err := n.db.HostsForScanning(time.Unix(0, 0), 100)
+	if err != nil {
+		t.Fatal(err)
+	}
+	testutil.Equal(t, "len(hosts)", 1, len(hosts))
+
+	now := types.CurrentTimestamp()
+	settings := proto2.HostSettings{
+		AcceptingContracts: true,
+	}
+	priceTable := proto3.HostPriceTable{
+		HostBlockHeight: 123,
+	}
+	// successful scan; should update settings and price table
+	scan1 := explorer.HostScan{
+		PublicKey:  hosts[0].PublicKey,
+		Success:    true,
+		Timestamp:  now,
+		NextScan:   now.Add(time.Hour),
+		Settings:   settings,
+		PriceTable: priceTable,
+	}
+
+	if err := n.db.AddHostScans([]explorer.HostScan{scan1}...); err != nil {
+		t.Fatal(err)
+	}
+
+	lastAnnouncement := n.tipBlock().Timestamp
+	host1 := explorer.Host{
+		PublicKey:              hosts[0].PublicKey,
+		NetAddress:             netAddr1,
+		KnownSince:             lastAnnouncement,
+		LastScan:               scan1.Timestamp,
+		LastScanSuccessful:     true,
+		LastAnnouncement:       lastAnnouncement,
+		NextScan:               scan1.NextScan,
+		TotalScans:             1,
+		SuccessfulInteractions: 1,
+		FailedInteractions:     0,
+		Settings:               settings,
+		PriceTable:             priceTable,
+	}
+	assertHost(hosts[0].PublicKey, host1)
+
+	now = types.CurrentTimestamp()
+	// unsuccessful scan
+	scan2 := explorer.HostScan{
+		PublicKey: hosts[0].PublicKey,
+		Success:   false,
+		Timestamp: now,
+		NextScan:  now.Add(time.Hour),
+		Error: func() *string {
+			x := "error"
+			return &x
+		}(),
+	}
+
+	if err := n.db.AddHostScans([]explorer.HostScan{scan2}...); err != nil {
+		t.Fatal(err)
+	}
+	// previous settings and price table should be preserved in case of failure
+	host1.LastScan = scan2.Timestamp
+	host1.NextScan = scan2.NextScan
+	host1.LastScanError = scan2.Error
+	host1.LastScanSuccessful = false
+	host1.TotalScans++
+	host1.FailedInteractions++
+
+	assertHost(hosts[0].PublicKey, host1)
+}
+
+func TestV2HostScan(t *testing.T) {
+	pk1 := types.GeneratePrivateKey()
+
+	n := newTestChain(t, true, nil)
+
+	assertHost := func(pubkey types.PublicKey, expected explorer.Host) {
+		hosts, err := n.db.QueryHosts(explorer.HostQuery{PublicKeys: []types.PublicKey{pubkey}}, explorer.HostSortPublicKey, explorer.HostSortAsc, 0, math.MaxInt64)
+		if err != nil {
+			t.Fatal(err)
+		}
+		testutil.Equal(t, "len(hosts)", 1, len(hosts))
+
+		testutil.Equal(t, "Host", expected, hosts[0])
+	}
+
+	// announce a host
+	const netAddr1 = "127.0.0.1:1234"
+	ha1 := chain.V2HostAnnouncement{{
+		Protocol: "http",
+		Address:  netAddr1,
+	}}
+
+	txn1 := types.V2Transaction{
+		Attestations: []types.Attestation{ha1.ToAttestation(n.tipState(), pk1)},
+	}
+
+	n.mineV2Transactions(t, txn1)
+
+	hosts, err := n.db.HostsForScanning(time.Unix(0, 0), 100)
+	if err != nil {
+		t.Fatal(err)
+	}
+	testutil.Equal(t, "len(hosts)", 1, len(hosts))
+
+	now := types.CurrentTimestamp()
+	settings := proto4.HostSettings{
+		ProtocolVersion:    [3]uint8{1, 2, 3},
+		AcceptingContracts: true,
+		Prices: proto4.HostPrices{
+			TipHeight: 123,
+		},
+	}
+	// successful scan; should update settings and price table
+	scan1 := explorer.HostScan{
+		PublicKey:  hosts[0].PublicKey,
+		Success:    true,
+		Timestamp:  now,
+		NextScan:   now.Add(time.Hour),
+		V2Settings: settings,
+	}
+
+	if err := n.db.AddHostScans([]explorer.HostScan{scan1}...); err != nil {
+		t.Fatal(err)
+	}
+
+	lastAnnouncement := n.tipBlock().Timestamp
+	host1 := explorer.Host{
+		V2:                     true,
+		PublicKey:              hosts[0].PublicKey,
+		V2NetAddresses:         []chain.NetAddress(ha1),
+		KnownSince:             lastAnnouncement,
+		LastScan:               scan1.Timestamp,
+		LastScanSuccessful:     true,
+		LastAnnouncement:       lastAnnouncement,
+		NextScan:               scan1.NextScan,
+		TotalScans:             1,
+		SuccessfulInteractions: 1,
+		FailedInteractions:     0,
+		V2Settings:             settings,
+	}
+	assertHost(hosts[0].PublicKey, host1)
+
+	now = types.CurrentTimestamp()
+	// unsuccessful scan
+	scan2 := explorer.HostScan{
+		PublicKey: hosts[0].PublicKey,
+		Success:   false,
+		Timestamp: now,
+		NextScan:  now.Add(time.Hour),
+		Error: func() *string {
+			x := "error"
+			return &x
+		}(),
+	}
+
+	if err := n.db.AddHostScans([]explorer.HostScan{scan2}...); err != nil {
+		t.Fatal(err)
+	}
+	// previous settings and price table should be preserved in case of failure
+	host1.LastScan = scan2.Timestamp
+	host1.NextScan = scan2.NextScan
+	host1.LastScanError = scan2.Error
+	host1.LastScanSuccessful = false
+	host1.TotalScans++
+	host1.FailedInteractions++
+
+	assertHost(hosts[0].PublicKey, host1)
+}
 
 func TestLastSuccessScan(t *testing.T) {
 	db, err := OpenDatabase(filepath.Join(t.TempDir(), "explored.sqlite3"), zaptest.NewLogger(t).Named("sqlite3"))
